Make it possible to completely ignore a test regardless if it passes.
The tailrecWithExplicitCompanionObjectDispatcher.kt test passes light
analysis test so the test wants to remove the `IGNORE_BACKEND: JVM`
directive. But the test does not work with JVM backend because it's not
compiled as a tailrec function.

This change makes it possible to completely ignore a test regardless of
whether it passes or not. Also with this change, directive
`IGNORE_LIGHT_ANALYSIS` means don't do anything for this test case.
Previously, it means automatically assume this test passes. And hence
if a directive disabling the test is present, this test should fail.